+++ piksi_multi_cpp/include/piksi_multi_cpp/sbp_callback_handler/utc_time_buffer.h
@@ -1,5 +1,5 @@
-#ifndef PIKSI_MULTI_CPP_SBP_CALLBACK_HANDLER_SBP_CALLBACK_HANDLER_RELAY_UTC_TIME_BUFFER_H_
-#define PIKSI_MULTI_CPP_SBP_CALLBACK_HANDLER_SBP_CALLBACK_HANDLER_RELAY_UTC_TIME_BUFFER_H_
+#ifndef PIKSI_MULTI_CPP_SBP_CALLBACK_HANDLER_UTC_TIME_BUFFER_H_
+#define PIKSI_MULTI_CPP_SBP_CALLBACK_HANDLER_UTC_TIME_BUFFER_H_
 
 #include <libsbp/sbp.h>
 #include <ros/ros.h>
@@ -27,4 +27,4 @@ class UtcTimeBuffer : public SBPCallbackHandler {
 
 }  // namespace piksi_multi_cpp
 
-#endif  // PIKSI_MULTI_CPP_SBP_CALLBACK_HANDLER_SBP_CALLBACK_HANDLER_RELAY_UTC_TIME_BUFFER_H_
+#endif  // PIKSI_MULTI_CPP_SBP_CALLBACK_HANDLER_UTC_TIME_BUFFER_H_
